Move additional methods to TimeUtils
