+++ app/src/main/java/org/jellyfin/androidtv/util/TimeUtils.java
@@ -1,10 +1,13 @@
 package org.jellyfin.androidtv.util;
 
+import android.text.format.DateFormat;
+
 import org.jellyfin.androidtv.R;
 import org.jellyfin.androidtv.TvApp;
 
 import java.util.Calendar;
 import java.util.Date;
+import java.util.Locale;
 import java.util.TimeZone;
 
 public class TimeUtils {
@@ -109,5 +112,48 @@ public class TimeUtils {
         }
 
         return convertedDate;
+    }
+
+    public static int numYears(Date start, Date end) {
+        Calendar calStart = Calendar.getInstance();
+        calStart.setTime(start);
+        Calendar calEnd = Calendar.getInstance();
+        calEnd.setTime(end);
+        return numYears(calStart, calEnd);
+    }
+
+    public static int numYears(Date start, Calendar end) {
+        Calendar calStart = Calendar.getInstance();
+        calStart.setTime(start);
+        return numYears(calStart, end);
+    }
+
+    public static int numYears(Calendar start, Calendar end) {
+        int age = end.get(Calendar.YEAR) - start.get(Calendar.YEAR);
+        if (end.get(Calendar.MONTH) < start.get(Calendar.MONTH)) {
+            age--;
+        } else if (end.get(Calendar.MONTH) == start.get(Calendar.MONTH)
+                && end.get(Calendar.DAY_OF_MONTH) < start.get(Calendar.DAY_OF_MONTH)) {
+            age--;
+        }
+        return age;
+    }
+
+    public static String getFriendlyDate(Date date) {
+        return getFriendlyDate(date, false);
+    }
+
+    public static String getFriendlyDate(Date date, boolean relative) {
+        Calendar cal = Calendar.getInstance();
+        cal.setTime(date);
+        Calendar now = Calendar.getInstance();
+        if (cal.get(Calendar.YEAR) == now.get(Calendar.YEAR)) {
+            if (cal.get(Calendar.DAY_OF_YEAR) == now.get(Calendar.DAY_OF_YEAR)) return TvApp.getApplication().getString(R.string.lbl_today);
+            if (cal.get(Calendar.DAY_OF_YEAR) == now.get(Calendar.DAY_OF_YEAR)+1) return TvApp.getApplication().getString(R.string.lbl_tomorrow);
+            if (cal.get(Calendar.DAY_OF_YEAR) < now.get(Calendar.DAY_OF_YEAR)+7 && cal.get(Calendar.DAY_OF_YEAR) > now.get(Calendar.DAY_OF_YEAR)) return cal.getDisplayName(Calendar.DAY_OF_WEEK, Calendar.LONG, Locale.getDefault());
+            if (relative) return String.format(TvApp.getApplication().getString(R.string.lbl_in_x_days), cal.get(Calendar.DAY_OF_YEAR) - now.get(Calendar.DAY_OF_YEAR));
+        }
+
+        return DateFormat.getDateFormat(TvApp.getApplication()).format(date);
     }
 }
